% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sidebar.R
\name{mod_sidebar_ui}
\alias{mod_sidebar_ui}
\alias{mod_sidebar_server}
\title{mod_sidebar_ui and mod_sidebar_server}
\usage{
mod_sidebar_ui(id)

mod_sidebar_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{Base à analyser, variables à descrire, options graphisme}

\item{output}{Tableau ou graphe}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
